+++ src/main/java/roomescape/reservation/ui/dto/request/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.ui.dto;
+package roomescape.reservation.ui.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
