+++ src/main/java/roomescape/reservation/ui/dto/request/CreateReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.ui.dto;
+package roomescape.reservation.ui.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
