+++ src/main/java/roomescape/reservation/ui/dto/request/ReservationsByfilterRequest.java
@@ -1,10 +1,10 @@
-package roomescape.reservation.ui.dto;
+package roomescape.reservation.ui.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 
 public record ReservationsByfilterRequest(
-        
+
         @NotNull
         Long themeId,
         @NotNull
