+++ src/main/java/roomescape/reservation/ui/dto/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.ui.dto;
+package roomescape.reservation.ui.dto.response;
 
 import java.time.LocalDate;
 import roomescape.member.ui.dto.MemberResponse.IdName;
