+++ src/test/java/roomescape/reservation/application/ReservationServiceTest.java
@@ -29,9 +29,9 @@ import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.ReservationTimeRepository;
 import roomescape.reservation.infrastructure.JdbcReservationRepository;
 import roomescape.reservation.infrastructure.JdbcReservationTimeRepository;
-import roomescape.reservation.ui.dto.AvailableReservationTimeRequest;
-import roomescape.reservation.ui.dto.AvailableReservationTimeResponse;
-import roomescape.reservation.ui.dto.CreateReservationRequest;
+import roomescape.reservation.ui.dto.request.AvailableReservationTimeRequest;
+import roomescape.reservation.ui.dto.request.CreateReservationRequest;
+import roomescape.reservation.ui.dto.response.AvailableReservationTimeResponse;
 import roomescape.theme.applcation.ThemeService;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeRepository;
